figured out how to print a long long integer!!!! now getting ready to implement unit testing

+++ libft/ft_itoa.c
@@ -6,7 +6,7 @@
 /*   By: mirivera <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/28 21:17:42 by mirivera          #+#    #+#             */
-/*   Updated: 2019/07/19 12:01:12 by mirivera         ###   ########.fr       */
+/*   Updated: 2019/07/22 18:01:17 by mirivera         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 ** Line 30 determines # of places in the string
 */
 
-char	*ft_itoa(int n)
+char	*ft_itoa(int64_t n)
 {
 	int64_t	length;
 	int64_t sign;
